Revert "Strip HTML from inspection description"
This reverts commit e6be154f

Inspection description is not supposed to contain HTML, and IntelliJ itself does not render HTML in the Problems tool window. Issues with HTML in inspection descriptions should be reported to wherever those inspections are coming from.

Stripping HTML breaks any properly formatted inspections that include anything that looks like an HTML tag, so I'm reverting it.

I might remove unescaping HTML entities later, but I'm not currently aware of it breaking anything, so it can stay.